requirements.txt
numpy>=1.24,<2.0
pandas>=2.0,<3.0
statsmodels>=0.14,<0.15
scikit-learn>=1.3,<2.0
xgboost>=2.0,<3.0
google-cloud-bigquery>=3.11,<4.0
google-auth>=2.22,<3.0
python-dateutil>=2.8,<3.0
